Start moving OS-specific Docker stuff to separate files
